expression: corrected a problem parsing the not operator
TEST: verify that a NOT operator is parsed correctly when it appears on
the right side of a binary operator.

This test initially failed because the compile number constant (at the
bottom of the expression compiler chain) saw the first character of the
NOT operator and returned a null data type since it didn't understand
this character.  The null data type was than passed up through the chain
until it reached the compile summation function (since the add operator
was used in the test expression), which rejected it incorrectly
reporting that a numerical expression was expected.

This problem was corrected by modifying the compile numeric operand
function to look for a NOT operator after checking for an open
parenthesis and before calling the compile numeric constant function.

To reduce some code duplication, the two functions were extracted from
the compile not function, which included the get not operator codes
function (calls the compiler get word operator codes function with the
not precedence level) and the compile not operand function (calls the
compile not function recursively for the operand of the not operand,
converts the operand to integer as needed, and returns the data type
from the call to the add selected code function.

The compile numeric operand function also calls the get not operator
codes, and if codes are returned, returns the data type of the call
to the compile not operand function.  This didn't eliminate all of the
duplication (these two lines are the same as in the compile not
function except these is also an else to call the compile equality
function), but eliminating this duplication would have been counter
productive as there would still be an if with its then statement.