recreator: separated interface from implementation
The recreator class was getting rather heavy.  To reduce the size of the
recreator header file, the implementation of the recreator class was
separated from the interface.

Instead of using the pointer to implementation (PIMPL) pattern used else
where (which was actually tried on another branch that was not pushed to
the official repository), the abstract base class pattern was used.

The recreator class was changed to a virtual base class containing only
pure virtual functions.  The original recreator class was changed to a
recreator implementation class with override of the virtual functions
and moved to a new recreator implementation header file.  All of the
recreator member functions were assigned to this new class.

To create an instance of the recreator, a new static create member
function was added to the abstract recreator class taking the same
arguments as the original recreator constructor.  This function returns
a unique pointer to the recreator so that it is automatically deleted.
The program unit class recreate line function was modified to use this
function to create the recreator instance.  Since this is no a pointer
to an instance, calling the operator function is a big klugey (need to
use the syntax "pointer->operator()()"), the function operator function
was renamed to recreate.

While converting the needed recreator member functions to virtual
functions, it was noticed that the recreate print function consisted of
only three calls to recreator functions (feature envy).  Therefore, this
functionality was moved to a new add command keyword recreate function,
which pushes the command keyword if the recreator stack is empty or
prepends the command keyword.  The allowed the empty function to be
removed (the stack can be checked directory) and the prepend keyword
function made private.  The prepend keyword was also used by the
recreate end function, but this function can also call the new add
command keyword function.  This change allowed two less virtual
functions.

These changes caused a problem with the coverage tests where it was
reported that the recreator class dynamic destructor was not covered
(the static destructor was being called).  This is a similar problem
that occurred with the error classes, and some coverage tests were added
to calls these functions.  Unfortunately, this same trick could not be
used with the recreator class since the compiler prevents creating an
instance of the abstract recreator class.  The only way to get this to
work was to declare the pure virtual functions as non-pure, add dummy
implementations for each of these functions (which goes against having
an abstract base class), and then calling each of the virtual functions
so that they are not reported as not being covered.  The decision was to
not do this and just let this function be reported as uncovered (just
now a little extra needs to be taken since 100% coverage will no longer
be reported).