compiler: corrected column problem with bad word following expression
TEST: verify that the compiler has the correct column when there is an
unrecognizable word where an operator is expected (after an operand).

The column is important so that the position of the error can be
reported correctly.  This test was initially failing where the column
that was reported was after the word instead of the beginning.

To get this test pass, a new word column was added to the compiler
class, which is set to the column member in the get alpha only word
function just before getting the characters of the word.  The get
column function was then modified to return the column member if the
word member is empty, otherwise the new word column is returned.