compiler: corrected problem with eqv operator
TEST: verify that the EQV operator is parsed correctly when it follows
an integer with no white space.

Normally an 'E' character after a number is treated as the beginning of
an exponent unless it is followed by another letter, in which case it
could be the start of an operator that begins with an 'E' (of which
there is only the EQV) operator.  Therefore the constant number compiler
does not treat with as a bad exponent but rather lets the caller handle
whether what follows is valid.

In this situation, the constant number compiler was setting an internal
'possible operator' flag that was accessible via an access function.
However, this feature was not being used, and should not have been
implemented as it turns out that is did not solve the issue correctly.

This test initially failed because the constant number compiler class
saw the 'E' after the constant was followed by another letter (a 'Q')
and so set its internal possible operator flag, but the expression
compiler did not check this flag.  Since the 'E' character was already
consumed, the expression compiler saw that the next character was the
'Q' and did not understand a 'QV' operator and so stopped parsing the
expression.

The compiler class needs to be mad aware of this character that was
consumed (so that it can be made the first character of a word).  The
possible operator flag was changed to a next character member, which is
initialized to zero (no character).  If there is an 'E' character
followed by a letter, it was removing the 'E' character from the number
string, but it did not do anything with it when the possible operator
flag was set.  This was changed to put this character into the new next
character member.  The set possible operator function was renamed to
remove exponent character to better identify what it does.

The expression compiler compile number constant was changed where after
checking if there is a negate operator, a check was added to see if
the constant number compiler has a next character and if it does to pass
it to the compiler instance to a new parse keyword function.

The new parse keyword function of the compiler class takes the first
character of the keyword as an argument.  It proceeds to get the
remaining alphabetic word, then sets its word member to the first
character plus the remaining word.  The word is then available when it
is requested for the next word operator.

Two of the constant number unit tests that checked the operator of the
possible operator function was updated to check the next character
function.