compiler: corrected column when there is a bad 'e' operator
TEST: verify that the compiler contains the correct column when there
is a word starting with an 'e' but not a valid operator.

For this situation, the compiling of the expression should stop at the
'e' character and the column should be at the 'e' character so an error
will be reported correctly (it's up to the caller to determine if this
is an error and to report the error accordingly).

This test initially failed because the column was at the character
following the 'e' character because the word column member in the
compiler class was set to the current column in the get alpha only word
function, but in this situation, this function started at the second
character of the word.

The test was made to pass by first moving the assignment of the word
column member to the callers of the get alpha only word function.  For
the parse keyword function, it is set to the current column minus one
to compensate for the first character (the 'e') that was already
consumed.