const-num: changed from pimpl to abstract interface base class pattern
Following the change of the Recreator class to an abstract interface
base class pattern to hide the implementation, the constant number
compiler class was also changed to this pattern.  This eliminates the
extra level of indirection from the main class through the pointer to
implementation to get to the implementation member function (though
similar to the indirection through a virtual table pointer, use of the
virtual table pointer is not explicit in the code).

Similarly, there is a static create function to create an instance of a
constant number compiler.  The function operator function was changed
to the compile function so that it is clearer then using the now verose
"->operator()()" to call this function (from a pointer).

Unlike with the recreator class, the implementation class was declared
right in the source file so that nobody will be able to use this class
directly.  The unit tests were updated accordingly.