expression: changed from pimpl to abstract interface base class pattern
Continuing the changes of the other classes to the abstract interface
base class pattern to hide the implementation, the expression compiler
class was changed to this pattern.

Similarly, a static create function was added to create an instance of a
expression compiler.  The function operator function was changed
to the compile function.