command: changed from pimpl to abstract interface base class pattern
Continuing the changes of the other classes to the abstract interface
base class pattern to hide the implementation, the command compiler
class was changed to this pattern.

Similarly, a static create function was added to create an instance of a
expression compiler.  The function operator function was changed
to the compile function.  The unit tests were updated accordingly.

OTHER: fixed a long line missed from the previous commit, and the
initializer for the equality codes member of the compiler class was
moved from the constructor to be a member initializer.