abs: parse function name, argument and closing parenthesis
TEST: verify that the argument of the ABS function and the closing
parenthesis is parsed to the end of the expression.

This test initially failed since up to now, the only valid operands were
numeric constants, the NOT operator and an opening parenthesis.  The ABS
keyword was not of these.

The minimum to get this test to pass was to add a check after looking
for a NOT operator to get the keyword from the compiler and see if it is
equal to "ABS" (if there was no word, than the empty word would not be
equal to "ABS").  If it is, a new compile function function is called.

This new function checks first clears the word in the compiler and then
checks if the peek next character is an open parenthesis.  If it is, the
compile parentheses function is called since it consumes the open
parenthesis character, skips white space, gets a numeric expression,
checks and gets the closing parenthesis character, and returns the data
type of the numeric expression.

The compile function function will need to do something different, but
this was enough to get this initial function test to pass.  (This tests
and the tests that will follow will help guide the design.)