codes: reorganized codes and code info classes
The operator codes base class will be needed for the functions, so it
was appropriate to rename it to simple the codes class, and it was moved
to a new codes source file.  The operator info class, which the codes
select function returns, was also renamed to info and put into the codes
class.