table: renamed the precedence classes to table
The precedence and precedence info classes contained information about
the operators, but these are the classes that are the logically place to
put information about the function also.  Therefore, these classes and
their source and header files were renamed to the table and table info
classes and files (which a similar to the classes in the IBCP).

The precedence level enumeration was moved out of the table class and
simply named the precedence enumeration.  This will removed some
unnecessary clutter when this enumeration name is used.