functions: implemented infrastructure for function codes
The names of functions will need be looked to get their associated
codes.  A case insensitive string to codes reference map was added to
the table info class for the numeric function codes (functions that
return numeric values used in numeric expressions).

An add numeric function data function was added to the table info class
that take a reference and a pointer to a keyword string.  This function
adds to the numeric function codes map, and the code values are added to
the code value to keyword vector.  The function codes are accessed from
a new numeric function codes function that takes a case insensitive
string.

A get numeric function codes function was added to the compiler class.
It gets the current word from the word member (assuming it is already
set).  The static table numeric function codes function is called, and
if codes are returned, the compiler word is cleared, and the codes are
returned.

The expression compiler compile function function was change from
comparing directly for the "ABS" string to calling the compiler get
numeric function codes function.  If no codes are return, the null data
type is returned.  The compiler clear word call was removed since this
is now done in the compiler get numeric function codes function.

A new template function code class was added, which takes a argument
type enumerator as a template argument.  This template class will be
used to define the function code instances.

A new multiple type function codes class was added based on the codes
class.  This class will be used to define the function codes instances
for functions that take either a double or integer returning the same.
These include the ABS and SGN functions.  For now it contains a single
double code (which was all that was needed to keep the tests passing).
The select function of this class is not called but required for the
code to compile.