abs: implemented compiling of the abs function
TEST: verify that an expression with an ABS function is compiled
correctly.

To make this test pass, the codes returned from the get numeric function
codes is put into a local variable.  If there are codes, the opening
parenthesis character is checked and consumed before calling the compile
parentheses function to compile the argument.  The add selected code is
then called to put the ABS function instruction into the program code.

This function was returning a double data type, but this was changed to
a null data type temporary since it is not actually being checked that;
and the compile numeric operand function was changed to just call the
compile function function without doing anything with the return value.
Though incorrect, the compile numeric constant is than called, but no
test checks this incorrect code.  This will be corrected as more
functionality is implemented.