abs: return double data type for abs function
TEST: verify that the abs function with a double argument returns a
double data type.

To get this test to pass, the add function code function was given a
data type return type and the codes info result data type value is
returned.  The compile function function was changed to return the
return value from the add function code.  The compile numeric operand
function was changed (back) to check the return value from the compile
function function and if not a null data type, the data type is
returned (otherwise the compile numeric constant function is called.