abs: implemented compiling of the integer abs function
TEST: verify that an ABS function with an integer argument is compiled
correctly.

To get this test to pass, an integer ABS code instance was added (with
null function pointers).  The multiple type function codes class was
given in integer code member and the constructor changed to initialize
this member.  An new integer argument type was added for integer
function codes.  The select function was changed to return the
appropriate info for the data type argument.

The expression compiler compile function function was changed to get the
return value from the compile parentheses function and pass this to the
add function code function.