abs: recreate an abs function with a double argument
TEST: verify that an ABS function with a double argument is recreated
correctly.

A new recreate function function was added, which calls a new recreate
function virtual function of the recreator class.  The recreate function
implementation function first moves the top string into a local operand
variable, get keyword for the function is appended to the top string
that is now empty (the function keyword is obtained by calling the get
operator keyword function), and the operand is appended with
parentheses.

This was sufficient for recreating the single argument ABS function and
will be sufficient for all of the math functions.  It will need to be
expanded and changed one multiple argument functions are added.

A pointer to the recreate function function replaces the null function
pointer in the abs double code instance.

Since the get operator keyword is now also used for functions, it was
appropriately renamed to get code keyword.