abs: recreate an abs function with an integer argument
TEST: verify that an ABS function with an integer argument is recreated
correctly.

The null function pointer in the abs integer code instance was replaced
with the recreate function function pointer, but this was not sufficient
to get the test to pass.  If was still failing because the ABS keyword
was not added to the code value to keyword map.  This was due to the
multiple type function codes code values function was not including the
code value for the integer code in the vector that is returned, so this
was added.