project: changed from generic integer to specific 32-bit integer type
An effort was made to change all specific types to the use of the C++
auto variable definition feature.  There were some uses of the 64-bit
integer type in the math operator execute functions (used to detect
overflows outside the 32-bit integer range).  This was accomplishes
using the type value initializer syntax (type name with the initializer
value in braces) that is similar to the function style cast but does not
permit narrowing conversions.

This syntax was also attempted for some of the C++ static casts used to
convert integer values to double, however, the compiler reports an error
that this is a narrowing conversion (for any integer type).  Since the
plain 'int' type could be 64 bits on a platform, and the intention here
is for integer values to be 32 bits, all int types were changed to the
32-bit specific integer type (executer stack integer value, constant
dictionary integer values, etc.).

With some math operations setting the top integer from a 64-bit integer
(after checking for an overflow, so will be within the range of a 32-bit
integer), a new set top integer from 64-bit integer function was added
to the executer class that casts the 64-bit argument to a 32-bit integer
before setting the top stack item.