abs: added check for underflow with integer argument
TEST: verify that an integer argument to the ABS function of the minimum
integer value produces an overflow run error.

Two changes were needed to get this test pass.  The first was to add a
check for the minimum integer value to the execute abs integer function.
This change was not sufficient to make the test pass (which was still
failing from an STL string exception, which was caused due to there not
being any error marker in the recreated line).

To get the error markers into the recreated line, mark error start and
end functions calls were added to the recreator implementation class
recreate function, with the start before the ABS function keyword is
added and the end just afterward.  Only the keyword itself is marked.

Since the maximum negative integer check was the same as in the execute
negate integer function, this check was extracted into a new check
negative integer overflow inline function into the overflow header file.