sgn: implemented compiling of the sgn function
TEST: verify that an expression with an SGN function is compiled
correctly.

All that was required to make this test pass was to add the code
instances for the sgn double and integer code instances (initially with
null function pointers) along with the sgn codes instance.  The SGN
function is like the ABS function in that it supports both a double and
an integer argument (a multi-type function).