sgn: implemented execute of the sgn function
TESTS: verify that the sgn function gives the correct results for zero,
positive and negate double and integer arguments.

Each test was added and made to pass one at a time.  Once complete, both
the execute sgn functions were nearly identical except for the type of
data being worked on.  A number of changes were made to eliminate this
duplication.

The executer set top double and integer functions were renamed to the
overloaded set top function name - to allow this function to be called
from a template function.  A template member top function was added to
get the top stack item, for which two specializations were implemented
for the double and 32-bit integer types (these functions call the top
double and top integer functions) - this function allows for the needed
function to be called from a template function.

An inline template execute sgn function was then added where the
template argument is the type passed from the execute sgn double and
integer functions.