sqr: implemented compiling of the sqr function
TEST: verify that an expression with an SQR function is compiled
correctly.

There is a single SQR code that takes a double argument.  There will be
a number of math functions that similarly only take a double argument.
For these functions, a new math function codes class was implemented
that contains a single code.  The select function of this class simply
returns the code with a double return data type.  The sqr code instance
was added (with null function pointers) along this the sqr codes
instance with its keyword.