sqr: handle integer arguments correctly
TEST: verify that an integer argument is converted to double before the
SQR function operators on it.

The SQR function like all of the other math class functions accept only
a double argument, so integer arguments require an internal convert to
integer code inserted into the program code before the function code.
This test initially fails (gives a garbage answer) because the execute
sqr function assumes that the top of the execution stack contains a
double value so an integer value is interpreted incorrectly.  Before
making this test pass, some refactoring was done.

Code was extracted from the compile function function to a new compile
function arguments functions, which checks for the opening parentheses,
calls the compile parentheses function to compile the argument and calls
the add function code to add the function code to the program.  This
function will need to be expanded as more functions are implement to
take into account different argument types, different number of
arguments, and optional arguments.

The compile parentheses function was incorrectly calling the compile
numeric expression function for a double expected data type.  This did
not causes a problem since the compile numeric expression function was
only checking that the expression data type was not null if the expected
data type was not null.  However, it intended that it would insert the
appropriate internal convert code if the expression data type was
different than a non-null expected data type, but this was not
implemented as it was not needed.

This convert code insertion was now needed for the SQR function, so the
compile numeric expression function was modified to insert a convert to
double code (by calling the convert to double function) if the expected
data type is double and the expression data type if not (an "expected
numeric expression is still thrown if the expression data type is null).

The compile parentheses function was changed to have an expected data
type argument, which is passed to the compile numeric expression
function.  A null data type was added to the compile numeric operand
caller - if a null data type is returned, the error should be caught by
a caller further up the calling chain - but it doesn't now since there
are no tests yet for this condition).

The compile function arguments function also needs to pass an expected
data type to the compile parentheses function, but the data type is
dependent on the function - a null data type for multi-type functions
(to accept any numeric type), and a double for math functions.  A new
virtual argument data type function was added to the function codes base
class.  Override functions were implemented in the multi-type and math
function codes classes to return these data types.  This function is
then called to pass the appropriate data type.