compiler: separated compile expression functions
The compile expression function of the compiler class and associated
compile function of the expression compiler helper class had a dual
purpose - one to get an expression for a specific data type (for example
a double argument to a function), and one to see if there was an
expression (for example the PRINT command).

The first instance required an expression and should throw an exception
if none was found, but the second instance did not require an expression
and would return a null data type if none was found.  This is confusing
from the caller as sometimes the return value is ignored and sometimes
it is not depending on the needs of the caller.  Because of this, this
functionality was separated into two functions.

The expression compiler compile function was renamed to the compile
expression function for better clarity before separating into two
functions.  The check for an expression error (expected numeric
expression) was moved from the compile numeric expression to the compile
expression function taking an expected data argument, which was
necessary so the compile expression function not taking an argument (but
returning the data type) and call it without an exception being throw.

The data type argument was removed from the callers of the compile
expression function, which included the compile print function and the
various unit tests that used the form the expected a data type return
value.

OTHER: several of the expression compile functions we renamed because
they had operator abbreviations in there name, which reduced the
readability of them.