expression: check for an expression after a open parenthesis
TEST: verify that an expected numeric expression error is produced when
there is no expression after an open parenthesis character.

This test initially failed because the incorrect "expected closing
parentheses" error was produced.  This was caused due to the compile
parentheses function calling the compile numeric expression function
that was no longing checking for no expression.  This functionality was
moved to the compile expression function.

To make this test pass, the compile parentheses function was changed to
call the compile expression function, which does check for no expression
and throws an error.  However, this function was not returning a data
type, so it was modified to return the data type (which is ignored by
the compile expression function in the compiler class).