int: implemented integer function
TESTS: verify that INT function expressions are compiled, recreated and
executed correctly.

To speed along develop since all of the infrastructure is already in
place for math functions, compiling, recreating and executing were
all implemented (though each test was added and made to pass one after
another).

There is a single INT code that takes a double argument.  The int code
instance was added along with the int codes instance with its keyword.
It was then given the recreate function pointer.  Finally, the execute
int function was implemented and its pointer added to the codes
instance.  The INT function is the equivalent to the standard floor math
function.  Though not needed for coverage, tests include both positive
and negative numbers with fractional parts close to and far from the
integer parts (.01 and .99) to verify that no rounding occurs.  Note
that for negative numbers, the result is below the integer value since
the INT function gives the nearest integer 'below' the argument.