fix: implemented fix function
TESTS: verify that FIX function expressions are compiled, recreated and
executed correctly.

All tests were put into the same test case and each was made to pass
one after another.

There is a single FIX code that takes a double argument.  The fix code
instance was added along with the fix codes instance with its keyword.
It was given the recreate function pointer and the execute fix function
was implemented and its pointer added to the codes instance.

The FIX function is the equivalent to the standard truncate math
function.  Though not needed for coverage, tests include both positive
and negative numbers with fractional parts close to and far from the
integer parts (.01 and .99) to verify that no rounding occurs.  The FIX
function, unlike the INT function returns the integer value nearest to
zero, which means that for negative numbers, the result is the same as
integer part of the number just like for positive numbers.