frac: implemented fractional function
TESTS: verify that FRAC function expressions are compiled, recreated and
executed correctly.

All tests were put into the same test case and each was made to pass
one after another.

There is a single FRAC code that takes a double argument.  The frac code
instance was added along with the frac codes instance with its keyword.
It was given the recreate function pointer and the execute frac function
was implemented and its pointer added to the codes instance.

The FRAC function doesn't existing in GW-Basic or QBasic but is a useful
function (inspiration came from HP calculations of the same function).
It can be computed by subtracting the result of the standard truncate
math function of the argument from the argument.