cos: implemented cosine function
TESTS: verify that the COS function expressions are compiled, recreated
and executed correctly.

All tests were put into the same test case and each was made to pass
one after another.

COS is a math function.  The cos code instance was added along with the
cos codes instance with its keyword and was given the recreate function
pointer and the execute cos function was implemented and its pointer
added to the codes instance using the standard cos math function.