sin: implemented sine function
TESTS: verify that the SIN function expressions are compiled, recreated
and executed correctly.

SIN is a math function.  The sin code instance was added along with the
sin codes instance with its keyword and was given the recreate function
pointer and the execute cos function was implemented and its pointer
added to the codes instance using the standard sin math function.