tan: implemented tangent function
TESTS: verify that the TAN function expressions are compiled, recreated
and executed correctly.

TAN is a math function.  The tan code instance was added along with the
tan codes instance with its keyword and was given the recreate function
pointer and the execute cos function was implemented and its pointer
added to the codes instance using the standard tan math function.

Note: The tangent function overflows at 90 degrees or one-half PI
radians, however, this cannot be reproduced since the exact value of
one-half PI cannot be passed to the tan function, so even though a large
value is returned, it does not actually overflow.