atn: implemented arctangent function
TESTS: verify that the ATN function expressions are compiled, recreated
and executed correctly.

ATN is a math function.  The atn code instance was added along with the
atn codes instance with its keyword and was given the recreate function
pointer and the execute atn function was implemented and its pointer
added to the codes instance using the standard atan math function.