log: implemented logarithm function
TESTS: verify that the LOG function expressions are compiled, recreated
and executed correctly.

LOG is a math function.  The log code instance was added along with the
log codes instance with its keyword and was given the recreate function
pointer and the execute log function was implemented and its pointer
added to the codes instance using the standard log math function.

The LOG function cannot have a non-positive argument, therefore there
are two additional execution tests with a zero argument and a negative
argument to check that the correct run error is produced.