exp: implemented base-e exponential function
TESTS: verify that EXP function expressions are compiled, recreated and
executed correctly.

EXP is a math function.  The exp code instance was added along with the
exp codes instance with its keyword and was given the recreate function
pointer and the execute exp function was implemented and its pointer
added to the codes instance using the standard exp math function.

The EXP function can underflow, but since the value just goes to zero
for very large negative arguments, no check for this condition is needed
(though a test was still added).  The EXP function can also overflow for
sufficiently large arguments, so a check was needed of the standard exp
function returns a huge value (the indication of an out of range value).
Since this check was identical to a check in the power double integer
class, this function was moved out of this class as a separate inline
function in the overflow header file, with the additional reference to
an executer instance argument.