cdbl: implemented compiling of the convert to double function
TESTS: verify that CDBL function is compiled to the end of the
expression, that code is added to the program code, and that the return
data type is integer.

The CDBL function is different from the math functions where its return
data type is opposite of its argument type (similar to the CINT
function).  So a new convert function codes class was added for this
function to hold the code.  The select function returns the code with
the double return data type.  The argument data type function returns
the integer data type.  This class will probably change when the CINT
function is implemented.

The cdbl code instance was added with null function pointers along with
the int codes instance with its keyword.  These instances along with the
convert functions codes class functions were sufficient to get these
first set of compile tests to pass.

The second test that verifies that the cdbl code was added to the
program code also verifies that no internal convert code is inserted
before the function code by using an argument expression that will
return and integer (sum of two integers) by checking the size of the
code allowing for the two constants, add code and cdbl code.