cdbl: implemented execution of the convert to double function
TEST: verify that the CDBL function is executed correctly.

The execution of the CDBL function is identical to the internal convert
to double function.  Therefore, the execute convert to double function
and the internal convert to double code instance was moved from the
math operators source file to the math functions source file.  The
pointer to this function was then put into the cdbl code instance.