expression: add internal convert to integer when needed
TEST: verify that an internal convert to integer code is added to the
program before a function requiring an integer (CDBL).

The CDBL function is the first function requiring an integer argument.
Up to now the convert numeric expression function was only adding an
internal convert to double code.  A check was added to convert to an
integer if the expected data type is integer.

The compile numeric expression function was also changing the data type
to be return to double when calling the convert to double data type, but
this was not necessary.  If a non-null expected data type was passed in,
then the caller doesn't use this return data type.  If a null expected
data type was passed in, then the convert functions are not called.