cint: implemented compiling of convert to integer function
TEST: verify that a cint function expression is compiled to the end.

The CINT is another convert function, but takes an integer argument and
returns a double.  First, the convert function codes class needed to be
modified since its constructor expected a function code with a double
argument data type.

A new constructor was added taking a function code with an integer
argument data type.  The type of the code member was changed to just a
reference to a code so that it can be assigned either function code type
(note that classes derived from the code class are only different in
their 'type' so that no programming errors are made assigning codes to
the various codes class instances).  These changes were sufficient to
get the test to pass.