cint: changed to expect a double argument type
TEST: verify that the CINT function expects a double argument type.

The argument for this test was the addition of two double constants.  If
a constant was used, the constant would have been changed to the desired
data type.  This test was initially failing because the function codes
class was hard-wired to return an integer argument data type, which
caused an internal convert to integer code (because CDBL has an integer
argument) to be added.

To get this test to pass, an argument data type member was added to the
function codes class, and each of the two constructors was modified to
set it to the correct type for the function code.  The argument data
type function was changed from returning the hard-coded integer data
type to returning the value of this new member.