cint: changed to have a double return type
TEST: verify that the CINT function returns double data type.

This test was initially failing because the convert function codes class
select function was hard-wired to return a double return data type
(because CDBL has a double return data type).

To get this test to pass, a return data type member was added to the
convert function codes class, and each of the two constructors was
modified to set it to the correct type for the convert function code.
The select function was changed from returning the hard-coded double
return data type to returning the value of this new member.